paddles_engine rules!
a macro for arbitrary user-generated rules

you define each rule in a stage: control, physics, collision, and
resources, in the order they're executed. each rule declares whether it
zips two output tables or filters them, corresponding to the actions
currently available in ConditionTables and Compose.

finally, optionally, the user can define if they want a predicate to run
on each value of the resulting output table ("foreach"), or a predicate
to run if at least one is true ("ifany").

see paddles.rs for an example